requirements.txt
pydantic
python-dotenv
python-telegram-bot
pydantic_settings
openai
aio_pika

fastapi
flake8
aiosqlite
httpx
uvicorn
pydantic_settings
greenlet
